Update to XXH3 for CRC_OPT

+++ src/CRC.h
@@ -4,5 +4,5 @@
 void CRC_Init();
 
 u32 CRC_Calculate_Strict( u32 crc, const void *buffer, u32 count );
-u32 CRC_Calculate( u32 crc, const void *buffer, u32 count );
-u32 CRC_CalculatePalette( u32 crc, const void *buffer, u32 count );
+u64 CRC_Calculate( u64 crc, const void *buffer, u32 count );
+u64 CRC_CalculatePalette( u64 crc, const void *buffer, u32 count );
